requirements.txt
anyio==4.11.0
astroid==2.15.8
attrs==25.4.0
beautifulsoup4==4.12.3
black==24.8.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.1.7
coverage==7.10.7
dill==0.4.0
exceptiongroup==1.3.0
fastapi==0.95.2
h11==0.16.0
idna==3.11
iniconfig==2.1.0
isort==5.13.2
lazy-object-proxy==1.12.0
lxml==5.3.0
mccabe==0.7.0
mypy_extensions==1.1.0
numpy==2.0.2
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
pip==25.3
platformdirs==4.4.0
pluggy==1.6.0
pydantic==1.10.24
pylint==2.17.7
PySocks==1.7.1
pytest==7.4.4
pytest-cov==4.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.31.0
selenium==4.36.0
setuptools==58.0.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
starlette==0.27.0
tomli==2.3.0
tomlkit==0.13.3
trio==0.31.0
trio-websocket==0.12.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.22.0
websocket-client==1.9.0
wrapt==1.17.3
wsproto==1.2.0